pyproject.toml
[tool.poetry]
name = "pymint"
version = "0.1.0"
description = "MINT is a language to describe Microfluidic Hardware Netlists. MINT is the name of the Microfluidic Netlist language used to describe microfluidic devices for Fluigi to place and route. Mint is a flavor of (MHDL) Microfluidic Hardware Description Language that can be used to represent Microfluidic Circuits."
authors = ["Radhakrishna Sanka <rkrishnasanka@gmail.com>"]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = "~3.8"
pip = "^20.2.2"
install = "^1.3.3"
antlr4-python3-runtime = "^4.8"
parchmint = "^0.1.2"

[tool.poetry.dev-dependencies]

[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "setuppy", path = "setup.py"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

